Działa wypożyczanie!

+++ src/main/java/com/example/jsfdemo/web/RentBean.java
@@ -80,8 +80,8 @@ public class RentBean implements Serializable {
 	public void loadSelectedCar(ActionEvent ae) {
 		String ids = FacesContext.getCurrentInstance().getExternalContext()
 				.getRequestParameterMap().get("carId").toString();
-		int id = Integer.parseInt(ids);
-		carId = (long) id;
+		Long id = Long.parseLong(ids);
+		carId = id;
 	}
 
 	public List<Car> getAllCars(){
